+++ .storybook/theme/theme.tsx
@@ -9,14 +9,35 @@ import light, { dark } from "./theme.definitions";
 const channel = addons.getChannel();
 
 export const ThemeDocsProvider: React.FC<any> = (props) => {
-  const [isDark, setDark] = useState();
+  const [isDark, setDark] = useState<boolean>();
+
+  useEffect(() => window.addEventListener("message", receiveMessage), []);
+
+  const receiveMessage = (message: MessageEvent) => {
+    switch (message.data?.type) {
+      case "change_dark_theme":
+        setDark(message.data.value);
+        break;
+      case "change_css":
+        const styleElement = document.createElement("style");
+        styleElement.innerHTML = message.data.value;
+        document.head.appendChild(styleElement);
+        break;
+      default:
+        break;
+    }
+  };
+
   useEffect(() => {
     channel.on(DARK_MODE_EVENT_NAME, setDark);
-
     return () => channel.removeListener(DARK_MODE_EVENT_NAME, setDark);
   }, [channel, setDark]);
 
-  return <DocsContainer {...props} theme={isDark ? dark : light} />;
+  return (
+    <ThemeProvider theme={isDark ? "dark" : "base"}>
+      <DocsContainer {...props} theme={isDark ? dark : light} />
+    </ThemeProvider>
+  );
 };
 
 export const ThemeRaruiProvider: React.FC<{ children: React.ReactNode }> = ({